requirements.txt
email-validator==2.2.0
fastapi==0.115.2
fastapi-cache2==0.2.2
httpx==0.27.2
motor==3.6.0
pydantic_settings==2.6.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
redis==5.1.1
uvicorn==0.32.0
